add editProfile endpoint
