+++ src/apis/accounts.js
@@ -38,6 +38,15 @@ export const apiMirrorAccount = async (address, signature, username) =>
     data: { address, signature, username }
   });
 
+  export const editProfile = async ({username, avatar, bio, fullname, authInfo}) =>
+  callYupApi({
+    url: `/accounts/edit-account/${username}`,
+    method: 'POST',
+    data: { avatar,
+      bio,
+      fullname,
+      ...authInfo}
+  });
 export const apiUploadProfileImage = async (imageData) =>
   callYupApi({
     url: '/accounts/account/profileImage',
